database-compare-back/requirements.txt
# Web框架
fastapi==0.109.0
uvicorn[standard]==0.27.0

# 数据库ORM
sqlalchemy==2.0.25
aiosqlite==0.19.0

# 数据验证
pydantic==2.5.3
pydantic-settings==2.1.0

# 数据库驱动
pymysql==1.1.0
oracledb==2.0.0
pymssql==2.2.11
psycopg2-binary==2.9.9
asyncpg==0.29.0
pyhive==0.7.0
thrift==0.16.0
thrift_sasl==0.4.3

# 加密
cryptography==41.0.7

# 日志
loguru==0.7.2

# 导出
openpyxl==3.1.2
jinja2==3.1.3

# 工具
python-multipart==0.0.6
python-dotenv==1.0.0
httpx==0.26.0
websockets==12.0

# 开发
pytest==7.4.4
pytest-asyncio==0.23.3
black==24.1.1
isort==5.13.2
